Make checkedTrenchesData update automatically from checkedTrenchesNames, and checkedTrenchItems reactive from checkedTrenchesData
