+++ src/components/TheControlTrenches.vue
@@ -23,7 +23,6 @@
           <li
             v-if="trenchName.includes(trenchGroupName)"
             class="list-group-item accordion"
-            @change="updateTrenchesDataWithSelectedTrench(trenchName)"
           >
             <input
               v-model="checkedTrenchesNames"
@@ -43,7 +42,7 @@
       <input
         :checked="areAllChecked"
         type="checkbox"
-        @click="toggleCheckAll(), fetchAllTrenchesData()"
+        @click="toggleCheckAll()"
       />
       <label class="px-1 m-0 text-bold" for="checkbox"> Check All </label>
     </div>
@@ -51,20 +50,17 @@
 </template>
 
 <script>
-import { apiFetchSurvey } from "@/services/ApiClient";
 import { mapActions, mapState, mapWritableState } from "pinia";
 import { useDataStore } from "@/stores/data";
 
 export default {
   data() {
     return {
-      checkedTrenchesVersions: {}, // use store ?
       isDisplayedArray: [],
     };
   },
-
   computed: {
-    ...mapState(useDataStore, ["projectTrenchesNames"]),
+    ...mapState(useDataStore, ["projectTrenchesNames", "checkedTrenchesItems"]),
     ...mapWritableState(useDataStore, [
       "checkedTrenchesNames",
       "checkedTrenchesData",
@@ -75,21 +71,26 @@ export default {
         .sort()
         .reverse();
     },
-    checkedTrenchesItems() {
-      let allItem = [];
-      Object.values(this.checkedTrenchesData).forEach((data) => {
-        allItem.push(...data);
-      });
-      return allItem;
-    },
     areAllChecked() {
       return (
         this.checkedTrenchesNames.length === this.projectTrenchesNames.length
       );
     },
   },
+  watch: {
+    checkedTrenchesNames(newValue, oldValue) {
+      // console.log("old", oldValue);
+      // console.log("new", newValue);
+      // console.log("======");
+
+      // Solution temporaire brutale. Il faudra comparer les longueurs de newValue et oldValue,
+      // et ne faire que le necessaire
+      this.checkedTrenchesData = {};
+      this.updateCheckedTrenchesData();
+    },
+  },
   methods: {
-    ...mapActions(useDataStore, ["setCheckedTrenchesItems", "selectedType"]),
+    ...mapActions(useDataStore, ["selectedType", "updateCheckedTrenchesData"]),
     toggleCheckAll: function () {
       // When the user click on check all
       if (this.areAllChecked) {
@@ -97,54 +98,6 @@ export default {
       } else {
         this.checkedTrenchesNames = [...this.projectTrenchesNames];
       }
-    },
-    updateTrenchesDataWithSelectedTrench: function (trench) {
-      if (
-        Object.prototype.hasOwnProperty.call(this.checkedTrenchesData, trench)
-      ) {
-        delete this.checkedTrenchesData[trench];
-      } else {
-        apiFetchSurvey(trench)
-          .then((response) => {
-            // prepare data to store in session in case of PUSH
-            this.checkedTrenchesData[trench] = response.data.surveys;
-            this.checkedTrenchesVersions[trench] = response.data.version;
-
-            // store in session in case of PUSH
-            sessionStorage.setItem(
-              "checkedTrenchesData",
-              JSON.stringify(this.checkedTrenchesData)
-            );
-            sessionStorage.setItem(
-              "checkedTrenchesVersions",
-              JSON.stringify(this.checkedTrenchesVersions)
-            );
-            this.setCheckedTrenchesItems(this.checkedTrenchesItems);
-          })
-          .catch(() => {});
-      }
-    },
-
-    fetchAllTrenchesData: function () {
-      this.checkedTrenchesNames.forEach((trench) => {
-        apiFetchSurvey(trench)
-          .then((response) => {
-            // prepare data to store in session in case of PUSH
-            this.checkedTrenchesData[trench] = response.data.surveys;
-            this.checkedTrenchesVersions[trench] = response.data.version;
-
-            // store in session in case of PUSH
-            sessionStorage.setItem(
-              "checkedTrenchesData",
-              JSON.stringify(this.checkedTrenchesData)
-            );
-            sessionStorage.setItem(
-              "checkedTrenchesVersions",
-              JSON.stringify(this.checkedTrenchesVersions)
-            );
-          })
-          .catch(() => {});
-      });
     },
   },
 };
